:hammer: Move bmson tests into src folder
